feat: 게임 진행 커맨드를 관리하는 기능 구현
